clip/requirement.txt
ipdb
joblib
cytoolz
cython
lz4
lmdb
msgpack-numpy
msgpack
toolz
simplejson
psutil
tensorboard
tqdm
easydict
pycocotools>=2.0.1
opencv-python
tensorboardX
torch==1.7.1
torchvision
ftfy
regex